Introduced AlbumServiceFacade +Impl. Introduced importSql to pre-populate Atrists and Countries tables in db. Preety much finished backend code, tested with postman. Moving slowly towards frontend.

+++ backend/MusicPublisherApp/src/main/java/pl/aIski/dto/SongCreationRequest.java
@@ -3,8 +3,9 @@ package pl.aIski.dto;
 import lombok.Data;
 
 import java.util.List;
+
 @Data
 public class SongCreationRequest {
     private String name;
-    private List<Integer> artistsIds;
+    private List<Long> artistsIds;
 }
